fix: replace identical code code with utils
